Make slack_webhook as non optional parameter
Initial the CLI argument slack_webhook was optional, even though other
options related to notificatin were required to be passed (Thanks to
Kirill for raising this point to me):

- app name
- app version
- notification context

I initially made the slack_webhook option to convey the fact that the
notification system is pluggable and you could instead optionally
integrate other providers like Discord or Microsoft teams. Making it
as non optional would indicate that this tool is tied to slack.

I discovered that this problem is solved by clap's ArgGroup, so I have
converted the solution to that. This also makes sure that we pass
atleast one parameter of the NotifyHook struct.